Addressed [0.5.2 All] Retain Zoom in structure #190
It would have been awesome if we could store viewport rect of the
structure and just restore it whenever the document is reopened.
However, its not always possible to do that. A document may be created
on a system with one resolution & structure canvas size and be opened on
another computer with another resolution and canvas size. While we can
retain zoom-level, we really cannot restore viewport rect exactly as is.