Fixed an issue that caused annotations filtered using StructureCanvasViewportFilterModel to not show up.
The issue was: bounding box of annotation items were always computed
using BoundingBoxEvaluator only for instantiated annotation items. That
meant that annotations filtered out using
StructureCanvasViewportFilterModel would never contribute to computation
of bounding box, that meant they would always remain outside the current
viewport.

To fix this issue, we now have a ModelAggregator that aggregates
bounding boxes of annotations in Structure. We use the aggregated
bounding box as the initialRect of the BoundingBoxEvaluator used in
StructureView.qml. This way bounding box of annotations are considered
whether or not they are visible.