Loading screenplays that were not created on the host computer takes a long time.
This is because ScreenplayEditor.qml has to compute geometries of all
scenes in order to setup its content-view (which will be a rather long
scroll area), atleast initially.

What we now do is get the ScreenplayAdapter to report only a treshold
number of scenes initially and show the document as quickly as possible.
But then we expand the adapter incrementally to match row count of the
underlying screenplay.