Revamping the Notebook tab.
- Previously notes were captured as a list of Note items. Now we have a
brand new Notes object which captures all sorts of Notes. Scene,
Character, Relationship and Structure now uses the new Notes class to
capture notes.
- Removed the old notebook view completely. We want to start fresh
anyway.
- Removed NotebookTabModel and replaced it with a brand new
NotebookModel, which is actually a tree model notes.
- Started reworking the Notebook tab UI by using a tree view to show all
available notes, clicking on an item should show its detail. This is
still a work in progress. It will be several commits before we complete
this.