Tab while typing character block should switch to Action paragraph.
https://www.scrite.io/index.php/forum/topic/cant-press-tab-to-get-back-
to-action/

While typing screenplay, 'tab' is a useful button to toggle between
action, character, dialogue, parenthesis etc. However, when it is in
'character', scrite offers the list of characters to select from (which
is good), and 'tab' button selects the character. 'Enter' button also
does the same job at that point, so Tab isn't needed for that job. The
trouble it is causing is that we can't jump to 'action' from 'character'
by pressing tab, as it will just select the first character in the list.
So the only way is to press ctr+1 (which also wasn't working at some
instances) or click on 'action' by mouse. Tab is a lot more convenient.