Documents created using this version should be openable in 0.7.12.
This version of Scrite brings to life a brand new workflow that involves
logging in to use certain parts of the app. The server infrastructure
needs to be tested before we can count on it. In the meantime, we don't
want users who create documents using this version to lose access to
their document. They should be able to open it in the previous version
of Scrite.