Create a device-id if QSysInfo::machineUniqueId() returns an empty byte- array.
It is possible that on some systems no unique machine Id can be queried.
In such cases we should create a unique device-id and store it along
with settings for use with the app. In-case the settings.ini is lost or
reset, then the app will have to be reactivated. That's all.