We now have a clang-format spec for Scrite.
The .clang-format file in root folder is based on what Qt uses for its
own project. It has been modified slightly to suit the purposes of this
project.

All the .h and .cpp files written for Scrite have now been formatted
using this. Code included from 3rdparty sources are not touched.