Print-Preview now renders a temporarily generated PDF file.
Until version 0.8.1, preview was generated by printing the screenplay
document into multiple images (one per page) using ImagePrinter.

ImagePrinter was using depricated APIs to redirect painting from a
QPrinter to one or more QImages. While the trick was neat, it was (a)
using depreciated APIs and (b) inaccurate, in the sense that page breaks
or even text alignments would be a bit off in rare cases.

Instead of using ImagePrinter, we now generate a temporary PDF export of
the screenplay and simply render that PDF in the UI instead. This
guarantees accurate preview display and also avoids us from having to
use depricated APIs.