Fixed an issue that caused screenplay editor tooltips to show in preview.
When preview mode is turned on, PdfView is layered on top of
ScreenplayEditor's contentView and other UI.

Since PdfView was not handling mouse click and hover events, they were
getting propagated to the items in ScreenplayEditor which are behind the
preview. This is obviously a bad idea. We don't want user to click on a
page in preview and trigger the opening of title page dialog for
example. That would be very confusing.