Updated the algorithm used to compute curved arrow connectors between index cards on the structure canvas.
There is a really good TypeScript library that can be used for efficient
computaton of curved arrow connectors between boxes.
Source: https://github.com/dragonman225/curved-arrows

While we were computing the curved connectors using a custom mechanism,
they were not generating optimum curves about 40% of the time. By
switching to curved-arrows from dragonman225, we are now able to render
beautiful curved connectors 100% of the time.

Since their implementation is in TypeScript, it had to be adapted to
JavaScript and then evaluated using QJSEngine from within the C++
StructureElementConnector Item class. JS evaluation can be slow,
especially when done at scale. So we delegate computation to a separate
thread to keep the main thread and render thread fluid and fast.