PDF reports are first generated into a temp folder and can then optionally be "downloaded".
Previously users would be forced to generate reports directly to PDF
files and then open them using a separate PDF viewer app.

Now the PDFs are first created in a temp folder, which the user will be
able to see right away within the Scrite app. If needed, those reports
can be saved/downloaded to the local file system.