Improved UX while resizing PDF pages in PdfView.qml
We render pdf-page images to suit the size of the rectangle on which it
is shown in the screen. That way the images don't appear "scaled". This
means that we have to reload PDF page images everytime there is a zoom
operation. But we don't want the page to become white and back during
every micro zoom operation, but just once after the zoom slider is
released.