Fixed multi-select scene resequencing on the Screenplay Editor.
If one or more scenes were range selected in the scene-list panel of the
Screenplay Editor and then were dragged to a new position, sometimes the
move wouldn't work and all the times the moved scenes wouldnt be
current. Also undo operation wouldnt work in Screenplay Editor after
move.

All of these are now fixed.