Finally screenplay resequencing works on Windows too.
- Dragging a selected item from scene-list panel was working. But its
possible to drag and item without selecting it first. In such cases, we
implicitly mark the dragged item as selected.

- Undo was not working because moving scenes on timeline or scene list
panel, will cause a bunch of position changes in StructureElement. These
changes would push their own undo commands on the stack, causing the
resequencing action to stay deep in the stack. Therefore the undo
operation wouldnt immediately happen until all the position-change-
commands are undone. This issue is fixed by ensuring that we collect
position changes as undoable actions only if they are triggered
explicitly by the user, and not as a part of other operations.