While modifying formatting options, the zoom level in preview is reset.
Within the settings dialog box, the font sizes in the preview depend on
zoom level of the screenplay editor. If the user has zoomed in/out in
the screenplay editor then the font sizes in preview appear too-big or
too-small to be of any judgemental value to the user.

We now retain the zoom level at 1 during the time the user is modifying
the formatting options and then reset it back to its original value as
soon as the dialog box is discarded.

This way font-sizes are in preview are consistent and natural.