Font size of synopsis text was not changing whenever screenplay editor was zoomed in/out.
It so happens that whenever the default font of TextArea is changed, the
syntax highlighter assigned on its document is not asked to highlight.

This causes FontSyntaxHighlighter to not reapply the zoomed font-size on
the text document. This in turn causes text in the synopsis editor to
appear out of sync with the rest of the text in screenplay editor.

Unfortunately, we don't get a signal whenever the default font on
QTextDocument is changed. So, we have to check in FontSyntaxHighlighter
if the QML item from which the document was fetched is a TextEdit or
TextArea component, in which case we have to hook to its fontChanged()
signal. Otherwise we have to hook to its widthChanged() and
heightChanged() signals. Which ever signal we hook to, we call
rehighlight() in the end so that the new font is applied on the
TextEdit/TextArea.