Introducing document vault.
Users would find all of their work/changes lost if the app crashed
before they had a chance to save it. This could happen for instance when
users either worked for a long time on a new document and forgot to save
it, OR if the user did save the file once, but has disabled auto-save
and the app OR OS crashed.

To address such unfortunate situations, we now have a document vault.
This vault stores unsaved files. It should ideally store all changes
until the most recent one, but sometimes it may be 1-2 seconds old.
Documents stored in the vault will get deleted if the user saves the
file explicitly. Otherwise users will have the option of restoring
document from the vault at anytime.