Form fields do not render text in selected fonts.
User feedback: "One more thing I'd like to share. In character sketch
and character interview tab user defined fonts do not work properly.
while selecting a box (physical description or goal) and start writing,
it is showing the fonts properly but as I deselect and click on the next
box ( background or internal conflict) the previous answered boxes are
disturbed and user defined fonts are not showing properly."

https://discord.com/channels/
867082699716689951/897442394348339210/936630486216880148