Added another checkbox in FinalDraftExporter dialog box to allow usage of Scrite fonts in FDX exports.
By default the Final Draft Exporter selects the first language-specific
font installed on the user's computer for encoding texts written using
that language. By turning on this option, Scrite will export text
fragments of different languages to use the same font that Scrite uses
to render them on screen and in PDF files.