New Feature: Attach a defining photo to each scene.
Some users have asked for a defining photo to be added to each scene so
that it can capture the core idea of the scene in question.

In a future commit, we will add a feature so that such photo(s) can be
optionally included in the exported PDF file so that it can also serve
as a story-boarding document.