New Feature: Rich Text Notes
Users wanted to be able to type notes in rich text. https://discord.com/
channels/867082699716689951/897442394348339210/934465277998030858

We use QuillJS to show a HTML/JavaScript based rich text editor within
the notebook.

At first, we did try to put together a rich text editor based on
TextArea and RichDocumentHandler (QObject), but it turned out to be a
pain. Instead of reinventing the wheel, we simply use a QuillJS based
editor in HTML.