Fixed a bug that caused text alignment to differ between on-screen and PDF versions of the screenplay.
This issue was caused because we were applying format changes in
response to onCheckedChanged signal in RadioButton. This is wrong
because the signal would get emitted even if the check state of radio
buttons get changed programmatically. We should ideally handle check
state changes only in response to user driven actions. To address this
we now handle onToggled to apply alignment changes to display and print
formats.