User Login Credentials Expire
1. The feedback we are getting from users is not helping us determine
why exactly the user credentials are getting discarded. To make it easy
for us to undertand this, we are displaying more information in the
error message that is displayed to the user.

2. Previously we would fetch new session tokens each time the app was
launched. This would cause any other concurrently running instance of
Scrite to logout, which then causes all instances to logout. Now we
explicitly request Scrite servers to not reset session tokens, if
another instance of Scrite is running on the same computer. This way we
can avoid force logout of users.