Changed UI of comments side panel to ditch TabBar and go with an icon based switch.
The tab-bar on the comments side panel was looking a bit too loaded. It
occupied too much space and was generally making the UI look heavy.

Instead the app now shows an icon based switcher on the left. It looks
lightly weight, does the job and makes the UI a bit more aesthetic.
(This was Surya's suggestion)