Scriptalay screenplays, templates and forms are now fetched using the REST API
Previously screenplay, template and form library from Scriptalay were
fetched as a compressed hex db file, which we would then uncompress and
showcase in the UI. Since 0.8, we have a dedicated web-service API for
all things Scrite, we might as well use that for fetching screenplays,
templates and forms. That way we don't have to rely on non-scrite.io
URLs for accessing information anymore.