Many users are reporting inflated UI on Windows.
It seems that in Windows 10 and later, when Settings > System > Scale
and layout > Change the size of text is set to 150% or more, the entire
UI bloats, not just the text.

Until we understand this better, lets disable DPI scaling on Windows.
Users can set custom font-point-size on Scrite if they want to zoom
texts. Atleast we can avoid a poor initial experience of Scrite on many
desktops.