BugFix: Hitting enter in auto-complete popup causes a wrong text to get picked
For example, if I type RUSS and the completion popup shows RUSSEL.
Hitting enter would complete RACHEL, which was the previously selected
item in the auto completion popup.

This was happening because the 'currentCompletion' property would only
update if current row changed and not if the filtered set of strings got
altered.