BugFix: White spaces in FDX files ignored while importing
QDomElement::setContent(QIODevice*) method seems to parse space strings
in DOM Elements as empty strings. This means when the following XML
element is parsed,

<Text> </Text>

We get QString(), instead of QString(" ")

This caused spaces to be gobbled up during FDX import.

The only way to fix this was to wrap a QXmlInputSource around the file-
device from which we are importing, and then as QDomDocument to use
QXmlSimpleReader to perform the actual parsing. This way all spaces are
properly parsed and the import is properly done.

As such QXmlInputSource and QXmlSimpleReader classes are deprecated in
Qt 5.15. But we will have to use them until an alternative is found. And
no QXmlStreamReader doesn't solve this problem.