BugFix: Using a dedicated timer to resolve screenplay editor blur
On M1 Macs, the editor blur wasn't disappearing even after scrolling
stabilized. Although, I am unable to reproduce the issue I have a hunch
that its probably because the timer is not furing after 400ms because we
are using execLater(). Trying out with a dedicated Timer {} to see if it
resolves the issue.