Changed implementation of execLater() method
- Previously we had an execLater() method in a C++ Application class,
which QML code would call to schedule a call for later execution. While
this worked for the most part, sometimes the context in which the
delayed-function-call should happen would get messed up.
- We have now implemented the whole execLater() method in implemented in
JavaScript and used in QML. This way the entire timer-code is created
and used in QML, therefore retaining call-contexts is natural.