BugFix: Window title doesnt show version number in file name.
Some writers prefer to include a version number in their file names. For
example, "My Screenplay v1.1.1.scrite" In such cases, the document name
we show in the title bar should be "My Screenplay v1.1.1" and not "My
Screenplay v1".

This can be resolved by using QFileInfo::completeBaseName() to fetch the
entire base name as opposed to the first base name (which
QFileInfo::baseName() returns).

But then this also means that we should use completeBaseName() in any
place we were using baseName() before. Otherwise we will have to deal
with inconsistent file names.