BugFix: If a paragraph has spelling mistakes, custom text & bg colors don't render with spell-check enabled.
Suppose that a paragraph of text has custom text and/or background
colors applied to one or more words in it. When spell check is enabled
and a paragraph of text has spelling mistakes, then the custom colors
are not rendered until all the spelling mistakes are fixed.

It appears that setting background color to light-red (as we used to do)
in syntax highlighter for misspelled words, causes all colors to
disappear in that paragraph. I am unable to reproduce this issue on a
separate sample Qt app; and I am unable to find out why its causing this
issue in Scrite.

Until we figure out why this is happening, we reapply background and
foreground colors whenever spelling mistakes are detected.