BugFix: Clicking on an item in scene-list-panel, freezes the app.
Screenplay Editor has a pull-out scene list panel. Users can quickly
navigate to a specific scene in the screenplay by clicking on an item in
that list. We had two problems in this

1. Clicking on an item in the list wouldn't provide instant feedback,
that the item was clicked on.
2. Switching to the selected item would take a long and during that time
the app freezes, showing no indication to the user that something is
happening.

Both these issues are now fixed.