BugFix: Clicking on an index card in structure or timeline freezes app
Clicking on an index card in the structure canvas or on the timeline app
wouldn't provide instant feedback, that the item was clicked on. This is
because the screenplay editor would freeze the app until it is able to
load the scene content. In large screenplays, this could take a while
and as such its not a good idea to keep the user unaware about what's
going on. We now show a busy indicator letting the user know that the
screenplay editor is loading stuff.