Optimisation: Using calc-scene-heights if heightHint is unavailable
We store height-hints for each scene along with the scene itself. This
lets the screenplay editor know how much screen height a scene is likely
to take once its fully loaded.

However, height-hints are not initially available. And even if they are
available, we discard them when they were calculated on another
computer.

In such cases, we would previously assume a hardcoded height hint of
400. Now we layout the scene text in an off-screen text document and
determine its height quickly (within 4-6ms). We then use that calculated
height as height-hint, until the delegate is fully loaded.

This leads to smoother scrolling experience on the screenplay editor.

Also, we now eliminate the use of Loader {} item to show the placeholder
image. Instead we dynamically create the Image item in memory and
destroy it when its no longer needed. This saves space for yet another
Loader {} item which would stuck around long after the scene delegate is
fully loaded.