BugFix: Loading backup in new-window reports user-login error
Even though user is logged in, we get an error message stating that a
document was locked and requires user login while opening a document
from a backup in new window.

This was because loading user-login from stored credentials would happen
a few milliseconds after the document is attempted to open.