BugFix: 50 Scene Issue
In odd cases, while number of scenes in the screenplay tab is 50 (for
example) - the number of scenes on the timeline in the same screenplay
within the structure tab is 51.

This gives an impression as if some scenes have disappeared from the
screenplay, when infact they have not. Switching to the screenplay tab
back from the structure tab seems to fix the issue.

https://discord.com/channels/
867082699716689951/1161217744923000882/1161217744923000882