Refactoring QML Code: Standardising all dialogs
We now have an AbstractDialog.qml that offers a standard dialog box
behaviour for all dialogs in Scrite. Eventually, we may add enter and
exit transitions like we had in modalDialog, but for now we will simply
use the built-in transitions that come with QtQuick Controls.

AboutDialog now extends AbstractDialog to offer its UI.