Refactoring QML Code: Title page gets its own dialog
Until now, title-page editor was shown within the application settings
dialog box. There are multiple issues with such an approach

1. We are mixing document settings with Application settings, which is a
bad idea.

2. The application settings dialog box, still works with the old
modalDialog construct, which we want to get rid of anyway.

From now on, title page fields are shown in its own TitlePageDialog
which is built on top of QtQuick Dialog.