Refactoring QML Code: MessageBox singleton
Previously we had showInformation() and askQuestion() functions in
main.qml which were used as global functions to show messages via the
old DialogOverlay or modalDialog id.

Since we are doing away with all leap-of-faith ids and also are moving
to standard QQC Dialogs, we now have a singleton QML item called
MessageBox which sports a function called showInformation(). This does
the same job as the previous function, but uses a dialog box from QQC.