Refactoring QML Code: Animated status texts shown using AnimatedTextOverlay now
Previously, we used a Loader with id statusText in main.qml and used
that from everywhere in the code to display animated status texts, for
scene/annotation copy.

Now, we use a singleton called AnimatedTextOverlay, which offers a
show() method - calling which an animated text is shown over the whole
UI.

This approach does away with using leap-of-faith access to the
statusText id in main.qml.