Refactoring QML Code: Hiding Theme page in Settings for now
Surya says that we should spend more time on designing the right way to
offer support for themes (or even tease dark mode). So, we are hiding
the theme page in Settings for now.

While using software-renderer QImageItem wouldn't render cover arts in
the home-screen. This issue has now been fixed.