Refactoring QML Code: VclText and VclLabel are different components.
Noticed several warnings of the form: "Both point size and pixel size
set. Using pixel size"

This was because previously VclText would set font.pointSize by default.
When it was instantiated and font.pixelSize would get used, this warning
would be written on the console.

Now, we have renamed the erstwhile VclText as VclLabel.

We have a new VclText component, which is pretty much VclLabel, except
that it doesnt specify default font.pointSize.