BugFix: We bundle OpenSSL 1.1.1w into the Linux AppImage bundle.
So far we were asking GNU Linux users of Scrite to download and install
OpenSSL 1.1.1 on their own. But now we bundle the SSL libraries into the
AppImage, include a complete copy of the OpenSSL source code and license
text in Scrite's source code and mention OpenSSL credits in
AboutDialog.qml.