NewFeature: We now show a warning sign for long scenes
Long scenes are those with word-count greater than a limit (250 by
default). Users can toggle display of this warning sign, or even change
the word limit.

Scrite is scene-centric, so its optimised for that. TextEdit isn't
optimal for editing large amounts of text, so its best if most of the
scenes in the screenplay aren't too long.