BugFix: Changed the way we look for changes to the currently open file.
First I tried tracking file changes using QFileSystemWatcher.
QFileSystemWatcher is supposed to emit fileChanged() signal anytime the
file at the stated path changes.

This seems to happen only when the file at the path was modified. It
doesn't happen, if the file was deleted and replaced with a new one by
the same name. This is because QFileSystemWatcher uses file-inode to
track changes, even if what we ask of it is to watch a path.

This is a deal-breaker for us, because there are instances where cloud-
sync-software (like OneDrive, GoogleDrive etc) delete the previously
synced file, only to replace it with an new one after sync. In such
cases the watcher doesnt emit the fileChanged() signal and the user
cannot be notified of the file being changed in the background by
another process.

Then, I tried using QFileSystemWatcher on the folder in which the file
existed and track changes to that folder instead. That wasn't reliable
as well. In some odd cases, the notification would never come. So, I had
to junk the idea of using QFileSystemWatcher all together.

For that reason, I now use a timer to periodically check for file
changes. Its a brute force way to monitor for changes to the file, but
it looks like we have no way out. If anybody has a better idea, please
do suggest.