[QML-Refactoring] By default add languages already available in the platform/OS
If the user has already added input methods / keyboard layouts for one
or more languages in the OS, then we might as well just add those
languages and leave the rest as is.

We also offer a "More Languages .." option in the Language menu to let
the user open the LanguageOptionsDialog, to add/remove/configure
languages.