[action-manager]: We now have a new way of working with actions.
Previously we were using Shortcut and ShortcutModelItem instances to
capture keyboard shortcuts, and populate a model. Multiple menu items,
tool buttons end up duplicating code and it was all a big mess.

Now, we will force usage of Action QML elements as the only way to
capture shortcuts and triggerrable actions. These can then be handled
using ActionHandlers or directly in the Action itself. Action objects
can be grouped into ActionManager instances.

Menus and Toolbars can be constructed from these ActionManagers.

All actions and their managers can easily be compiled into a list-model,
which is the ActionsModel. This is what can now be used to show
shortcuts in one place. This can also be used to offer a single page in
the Settings dialog box where keyboard shortcuts can be easily remapped.