[runtime] Breaking apart utility functions in Application class. [WIP]
The Application class had become a dump yard for functions that couldn't
go any specific place.

In this commit, they are distributed into a bunch of utility classes
that are accessible as singletons in C++ and in QML.

Makes the code easier to read, and stress-free to manage and extend in
the future.