[language-support] Language change shortcuts now work in VclDialog also
- Modal dialog boxes would naturally cause all language Action objects
in ActionHub.languageOptions to get disabled when the dialog box is
open. This is expected and documented behaviour in Qt, and there is no
way to selectively turn it off.

- In VclDialog, we now have a handleLanguageShortcuts property, which is
false by default. When set to true, the dialog box automatically creates
new Shortcut {} items for handling language switch whenever the dialog
box is visible.

- This property is set to true in TitlePageDialog,
RenameCharacterDialog, UserAccountDialog, ExportConfigurationDialog and
ReportConfigurationDialog. Because, we don't need it in any other dialog
as of now.