[screenplayeditor] BugFix: Changing the way we handle jump and scroll scenarios [WIP]
Previously, we would handle jump to previous scene, next scene, first
scene & last scene, along with scroll into previous and next scenes
within the TextEdit of the scene-content editor, and then cascade those
signals across to the list view and have that handle these.

The problem with this was that we couldn't get the keyboard shortcuts
for such actions to become configurable.

Now, we capture these actions as separate Action items in ActionHub, and
handle them via ActionHandlers in appropriate places.