[settings] Added config option to edit scene loading interval
Scene Loading Interval is the time in ms after which a high resolution
of the scene will be loaded while scrolling rapidly. A higher value will
keep scrolling smooth, but a lower value will cause content to load
faster. The default value is 250ms. This can be any value between 50 and
1000ms