[screenplayeditor] Ctrl+Alt+0 is editSceneNumber
Previously it was Ctrl+9. This seemed odd, because Ctrl+0 was mapped to
scene heading, it seemed logical to map Alt+0 to scene number, since
they are all related to scene heading area itself. But then Alt+0 is
already mapped to sidePanelVisibility, so we now use Ctrl+Alt+0 for
editSceneNumber action.