[paginator] Building a new paginator module
The paginator we currently have is very slow, and causes the UI to
freeze as the user is typing content. This makes the entire typing
experience poor, and also the pagination calculations inaccurate across
various modules.

For instance, the page count reported in screenplay editor is distinct
from what's reported in statistics report. We need a way to make this
ridiculously consistent across all modules within the app, even if the
actual page count in the PDF may vary (due to page breaks introduced
from MORE/CONTD, and the presense of act breaks, custom page breaks,
synopsis, character tags etc).

While ScreenplayTextDocument will continue to be used for PDF and ODT
generation, the new ScreenplayPaginator will be used for realtime
pagination and for estimation of scene & page lengths across the board.