[statisticsreport] Updated StatisticsReport to use the new ScreenplayPaginator for calculating pixel, page and time metrics.
Previously the StatisticsReport was using a different method for
evaluating pixel, page and time metrics in the screenplay than the one
we use for reporting metrics in the screenplay editor. This caused
discrepancies in many things, the most obvious being page count itself.
By using the same paginator in both places, we have now eliminated these
discrepancies.