[paginator] Auto adjust the sync-interval based on the time taken to paginate.
Pagination is now performed in a background thread. Changes made to the
document are batch updated at once in a background thread. The batch-
update interval is 500ms by default.

But if a particular document took more than that time, then the update
interval is rounded upwards to the nearest multiple of 500ms. And when
the time taken drops, it drops down up to 500ms.

This is done to ensure that the worker is not unncessarily burdened with
updates frequently.