[tooltips] Using a custom tooltip popup instead of ToolTip attached property to show tooltips
Why?

It so happens that built-in tooltip with ToolTip attached property
sometimes causes the tooltip to stack under the item against which the
tooltip is shown. That totally defeats the purpose. Instead, we now make
use of ToolTipPopup to show a custom tooltip popup, within which the
tooltip text is shown.