[commandcenter] Introducing command center
Just hit Ctrl+/ to fire up a popup which shows available application
commands. You can type a few characters to filter in that list, use
arrow keys to navigate, and hit return key to trigger the selected
command. It makes it easy to find actions in the app and trigger them.