[timelineview]: Refactored ScreenplayTracks model, and the rendering view item
Previously we would construct data in the ScreenplayTracks C++ model as
QVariantMap. This works, is probably efficient, except it's not
understandable by someone else.

We now return ScreenplayTrack struct/gadget instances from the model,
each of which is a name and a list of ScreenplayTrackItem items. Each
ScreenplayTrackItem is defined by startIndex, endIndex & name.

This makes the code easy to understand and maintain.