[screenplayeditor]: Fixed a bug that caused formal tags to not get checked properly.
Because we were not explicitly requiring the index property (in
StructureGroupsMenu's internal ListView that displays the formal tags),
the value of the index property would be derived from the parent
context, which is the ScreenplayElementListView in this case. This would
case the first element to always select the first tag, the second
element the second tag and so on.