[screenplayeditor]: Model for a potential tree view in the scene list panel
So far, we have only been showing a list view on the side panel. The
view is flexible, just that it doesnt help with visualizing the
hierarchy of the scenes in the screenplay. There has been a pending user
request for a while to show a tree view in there. With this commit, we
just add a ScreenplayTreeAdapter, which is eventually going to be
upgraded to offer a full blown tree view. For now, its just a model and
its not used in the app.