[screenplayeditor] Bugfix: Clicking on website link in title-page would show an error
If the website URL was relative, then Qt.openUrlExternally() would try
to open that link from within the app, which would show up a OS specific
dialog box asking for user to let them know what browser to use for a
qrc scheme.

We now check for the validity of the URL before trying to open it
externally, otherwise we simply show the title page dialog box.

Same with email as well.