[LanguageOptionsPage] Bugfix: Undefined language
When a language is removed, for a brief few milliseconds there is no
language specified. At this point, the is forced to show a random
language and can cause the QMLEngine to report errors or just give up.

We have to factor this possibility and allow the UI to gracefully handle
this.