[core]: Introducing DelayedProperty attached property
Consider a view/repeater scenario like this

Repeater {
   model: someObject.counter
   Item { ... }
}

If the counter value changes rapidly within a short duration, say 100ms,
then the repeater is obligated to create/destroy delegates for each
change instantly. This can add needless load to the QML engine.

Wouldn't it be nice to delay such a change?

Repeater {
   model: DelayedProperty.value

   DelayedProperty.watch: someObject.counter

   Item { ... }
}

The DelayedProperty attached object has a property called watch. Every
change to counter is accepted to the watch property immediately, and it
internally schedules a timer for 100ms by default, but you can configure
the delay using DelayedProperty.delay. After timeout, it updates its own
DelayedProperty.value, and the Repeater can create/destroy delegates all
at once.

Its also possible to use DelayedProperty.initial to a value, so that
whenver value of watch changes, value is instantly set to initial once
and then to the watched value after the timeout. This is useful in
scenarios where we want the first change to instantly reset the Repeater
delegates, but all other subsequent changes to bundle into a 100ms (or
delay) batch.

NOTE: This attached property is in addition to DelayedPropertyBinder
that we already have had for a while now. Both do the same work, except
one acts as a attached property and another as a standard QtObject.