[paginator] Added computation of scene and cursor offsets.
The paginator now reports along with each ScreenplayPaginatorRecord, the
cursor, pixel, page and time offset from which it starts within the
document.

This information is especially useful to evaluate relative offset of a
cursor within a record. The idea is to use this to eventually display an
actual timeline on the timeline in the structure tab. At the very least
we can show a cursor line within the current block indicating the
relative position of the text-cursor within a scene on the timeline.