[structurecanvas] Bugfix: Crash while deleting index cards in a selection.
This was caused due to a buggy GarbageCollector.

The GarbageCollector was built in an era where it was important to delay
the deletion objects to allow for the UI to adjust itself. I don't think
we need it anymore. So, we simply get GarbageCollector::add() method to
delete objects using qDeleteLater()